Attempt for fwrite accuracy to pass windows.

+++ src/fwrite.c
@@ -89,8 +89,8 @@ static inline void writeNumeric(double x, char **thisCh)
   } else {
     if (x < 0.0) { *ch++ = '-'; x = -x; }  // and we're done on sign, already written. no need to pass back sign
     int exp = (int)floor(log10(x));
-    unsigned long long l = (unsigned long long)((long double)x * powl(10, NUM_SF-exp));
-    // TODO?: use lookup table like base R? .................... ^^^^
+    unsigned long long l = (unsigned long long)(((long double)x) * (long double)powl(10, NUM_SF-exp));
+    // TODO?: use lookup table like base R? ................................... ^^^^
     //        here in fwrite it might make a difference whereas in base R other very
     //        significant write.table inefficiency dominates.
     // long double needed for 1729.9 to ensure 1e-310 doesn't write as 0.
